# Optimize load_py for memory and speed