This PR implements several strategies to improve the memory footprint and speed performance of `load_py` while maintaining backwards compatibility:

- The Zip archive streams are not opened and read until the moment they are needed.
- The Zip archive streams are read in the order that they are stored in the file to reduce random file seeks.
- Avoid holding intermediate tensors in memory before they are copied to the target module.
- Avoid intermediate copies of tensors during read, if possible.
- Any intermediate tensors or streams are allocated and disposed of as soon as possible.

## Test

These tests load the llama-2-7B-chat pickled model on a MacBook Air, M2 processor, with 24GB of memory and SSD.

### CPU Model

#### Baseline - 41 secs

<img width="202" alt="cpu-baseline-41-secs" src="https://github.com/user-attachments/assets/c62590ed-9afb-4ea3-aeff-13f816c3bfef">

#### With Optimizations - 9.6 secs

<img width="206" alt="cpu-optimized-9 6-secs" src="https://github.com/user-attachments/assets/b6959f01-846b-4ceb-8e40-05a0474cf018">

### MPS Model

#### Before - 41 secs

<img width="207" alt="mps-baseline-41-secs" src="https://github.com/user-attachments/assets/f84ebb5f-eea5-434a-a360-704fe3dc13b4">

#### After - 30 secs

<img width="204" alt="mps-optimized-30secs" src="https://github.com/user-attachments/assets/4d90ec84-4d46-4b4a-b2ea-c3d05460c254">