@shaltielshmid 2024-08-07
Looks great! Thank you @ejhg . 
I'll merge the changes in later today and then publish a new version to NuGet.