@shaltielshmid 2024-08-07
Thanks for your contributions @ejhg. I just uploaded version 1.4.2 to NuGet, should be available shortly.