@ejhg 2024-08-07
I don't think I mentioned this:

I feel there may still be room for improvement somewhere (I feel that the stream reading still has too much overhead on the OS side), but with this change, the `load_py` is running pretty much at "bare metal" efficiency (since the stream reading happens as close as possible to the handoff to Torch).  With this change, `load_py` is way faster and has less overhead than Python for the workload I tested (loading llama 7B).

Additional personal commentary:

I think that the unpickling implementation in PyBridge is well encapsulated.  Once I figured out what was going on, I realized that the code that handles the .net unzipping and stream reading and tensor loading is completely separate from the "python internals".

In my opinion, I think TorchSharp should adopt this implementation as the preferred method to read and write weights.  I think Python interoperatiblity should be a priority.

In the TorchSharp documentation, they mention that pickles are a "complex format", but now that `weights_only=True` is becoming the default option in `torch.load` (due to security issues), that can be used to limit the expected scope of  PyBridge's load_py implementation.  In other words, `load_py` is already doing the job of `torch.load(weights_only=True)` as far as I can tell.