@shaltielshmid 2024-08-08
Great points! I've also noticed a shift to using [safetensors](https://github.com/huggingface/safetensors), an alternative easy (and safe) tensor format. There was a discussion to add the `load_py` directly into TorchSharp, but we decided against it for a few reasons. Therefore, I created this library as an alternative. 

You are more than welcome to voice your thoughts on the topic, and we can open up the discussion again: https://github.com/dotnet/TorchSharp/discussions/1231.